+++ packages/sw-lib/test/sw/caching-strategies.js
@@ -23,7 +23,7 @@ describe('Test caching strategies.', function() {
   ];
   strategies.forEach((strategy) => {
     it(`should have '${strategy}' available`, function() {
-      expect(goog.swlib[strategy]).to.exist;
+      expect(goog.swlib.strategies[strategy]).to.exist;
     });
 
     const badArguments = [
@@ -43,20 +43,20 @@ describe('Test caching strategies.', function() {
     });
 
     it(`should return a Handler when '${strategy}' is instantiated without arguments`, function() {
-      const handler = goog.swlib[strategy]();
+      const handler = goog.swlib.strategies[strategy]();
       expect(handler.handle).to.exist;
       expect(handler.requestWrapper).to.exist;
     });
 
     it(`should return a Handler when '${strategy}' is instantiated with empty object`, function() {
-      const handler = goog.swlib[strategy]({});
+      const handler = goog.swlib.strategies[strategy]({});
       expect(handler.handle).to.exist;
       expect(handler.requestWrapper).to.exist;
     });
 
     it(`should return a Handler when '${strategy}' is instantiated with cacheName`, function() {
       const CACHE_NAME = 'hello-world-' + Date.now();
-      const handler = goog.swlib[strategy]({
+      const handler = goog.swlib.strategies[strategy]({
         cacheName: CACHE_NAME,
       });
       expect(handler.handle).to.exist;
@@ -69,7 +69,7 @@ describe('Test caching strategies.', function() {
         maxEntries: 10,
         maxAgeSeconds: 60 * 60,
       };
-      const handler = goog.swlib[strategy]({
+      const handler = goog.swlib.strategies[strategy]({
         cacheExpiration: CACHE_EXPIRATION,
       });
       expect(handler.handle).to.exist;
@@ -83,7 +83,7 @@ describe('Test caching strategies.', function() {
       const BROADCAST_CACHE_UPDATE = {
         channelName: CHANNEL_NAME,
       };
-      const handler = goog.swlib[strategy]({
+      const handler = goog.swlib.strategies[strategy]({
         broadcastCacheUpdate: BROADCAST_CACHE_UPDATE,
       });
       expect(handler.handle).to.exist;
@@ -100,7 +100,7 @@ describe('Test caching strategies.', function() {
           'Example-Header-2': 'Header-Value-2',
         },
       };
-      const handler = goog.swlib[strategy]({
+      const handler = goog.swlib.strategies[strategy]({
         cacheableResponse: CACHEABLE_RESPONSE_OPTIONS,
       });
       expect(handler.handle).to.exist;
